Add jmath, jresource, jsecurity; Add win32 directory provide support to VC2010; Add modules now use j<module>lib.h with all headers from module; Fix some bugs

+++ jcommon/include/jbitstream.h
@@ -63,7 +63,7 @@ class BitStream : public virtual jcommon::Object{
 		 * \brief
 		 *
 		 */
-		unsigned long long GetBits(int size);
+		uint64_t GetBits(int size);
 		
 		/**
 		 * \brief
