+++ jio/include/jmemoryinputstream.h
@@ -82,7 +82,7 @@ class MemoryInputStream : public jio::InputStream{
 		 * \brief
 		 * 
 		 */
-		virtual int Read();
+		virtual int64_t Read();
 
 		/**
 		 * \brief
